requirements.txt
mcp>=1.0.0
anyio>=4.0.0
Pillow>=10.0.0
